move old playground, change biome settings, clean up package/types, format
