+++ examples/withPlugins/.vscode/extensions.json
@@ -1,0 +1,4 @@
+{
+	"recommendations": ["astro-build.astro-vscode"],
+	"unwantedRecommendations": []
+}
