+++ examples/withPlugins/integration/plugin.ts
