+++ examples/withPlugins/integration/virtual.d.ts
@@ -1,0 +1,3 @@
+declare module "virtual:astro-integration-kit-playground/config" {
+	export default { foo: string };
+}
