+++ examples/withPlugins/src/components/Card.astro
