+++ examples/withPlugins/src/env.d.ts
@@ -1,3 +1,4 @@
+/// <reference path="../.astro/types.d.ts" />
 /// <reference path='../.astro/types.d.ts' />
 /// <reference types='../.astro/test-format.d.ts' />
 /// <reference types='../.astro/test-integration.d.ts' />
