+++ examples/withPlugins/src/pages/virtual-module.astro
