+++ examples/withPlugins/tailwind.config.mjs
