+++ examples/withPlugins/tsconfig.json
@@ -1,0 +1,3 @@
+{
+	"extends": "astro/tsconfigs/strictest"
+}
