+++ playground/package.json
@@ -2,24 +2,14 @@
 	"name": "playground",
 	"type": "module",
 	"version": "0.0.1",
-	"private": true,
 	"scripts": {
 		"dev": "astro dev",
-		"start": "astro dev",
-		"build": "astro check && astro build",
+		"build": "astro build",
 		"preview": "astro preview",
 		"astro": "astro"
 	},
 	"dependencies": {
-		"@astrojs/tailwind": "^5.1.0",
-		"astro": "^4.12.1",
-		"astro-integration-kit": "workspace:*",
-		"tailwindcss": "^3.4.3"
-	},
-	"devDependencies": {
-		"@astrojs/check": "^0.8.2",
-		"@types/node": "^20.12.12",
-		"typescript": "^5.4.5",
-		"vite-plugin-pwa": "^0.19.8"
+		"astro": "^5.12.4",
+		"astro-integration-kit": "workspace:*"
 	}
 }
